Add crc.xml converter kom plugin method.
This now makes it to where now the kom plugins does the actual crc.xml conversion stuff, the KOMStream.ConvertCRC does the things needed to invoke it properly for when packing to specific kom versions without issues. Now all we just need to do is finish the crc.xml updater stuff which probably should also call plugin methods that does it all.